requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
python-docx==0.8.11
markdown==3.4.4
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.24.3
sentence-transformers==2.2.2
chromadb==0.4.6
openai==0.27.8